+++ src/modules/assembler-globals.hpp
@@ -21,7 +21,7 @@ public:
   Globals() = default;
   ~Globals() = default;
 
-  std::vector<uint32_t> serialize();
+  std::vector<uint64_t> serialize();
 };
 
 void parseGlobalsSection(Globals &globals, webassembly_t::global_section_t &global_section);
